SONOPAController/requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.9
Tempita==0.5.2
Werkzeug==0.10.4
blinker==1.3
decorator==3.4.2
itsdangerous==0.24
jsonschema==2.4.0
mysql==0.0.1
numpy==1.9.2
pbr==0.11.0
requests==2.7.0
six==1.9.0
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
wsgiref==0.1.2
